# backend/solarhub-nestjs/swagger-output.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/shop/store/products": {
      "get": {
        "operationId": "StoreCatalogController_getProducts",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Store Catalog (B2C)"
        ]
      }
    },
    "/api/v1/shop/store/search": {
      "get": {
        "operationId": "StoreCatalogController_searchProducts",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Store Catalog (B2C)"
        ]
      }
    },
...
    "/api/v1/shop/store/companies/{id}/products": {
      "get": {
        "operationId": "StoreCatalogController_getCompanyProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Store Catalog (B2C)"
        ]
      }
    },
    "/api/v1/shop/store/categories/{type}/{id}/products": {
      "get": {
        "operationId": "StoreCatalogController_getCategoryProducts",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Store Catalog (B2C)"
        ]
      }
    },
...
  },
  "info": {
    "title": "SolarHub API",
    "description": "The SolarHub Backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
  }
}